==> packages/kptn/kptn-0.1.4.tar.gz/kptn-0.1.4/example/nibrs/tasks/compute_weights/06_Weight_Checks.Rmd <==
---
title: "Weights Checks"
author: "JD Bunker"
date: "07/06/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
#Note (16Jun2023): updating to include JD/MSA/FO checks
```

## R Markdown

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

## Purpose

The purpose of this program is to compile weight checks from national, region, state, judicial district, MSA, field office, tribal, and university weights

* NIBRS-reporting LEAs with missing weights are marked red in all the tables. Each of these LEAs should get a nonzero weight during calibration and thus missing weights implies an issue during the calibration process. If nonzero weights exist the estimation process should be halted as the estimates will not be correct. A review of how the LEA is identified for purposes of estimation should be reviewed to ensure it should receive a weight.  
* Weights greater than 20 in all tables are marked red in all the tables. This suggests that a single NIBRS-reporting agency represents 20+ NIBRS-eligible agencies and may yield unreliable estimates. If this occurs in a domain, review the coverage rate for the domain. If the coverage is low, this result is expected and not an indication of a larger issue. 
* In the national and regional tables, weights less than 0.9 are marked red and those that are at least 0.9 and less than 1 are marked yellow. During calibration of national- and region-level weights, emphasis was placed on fitting all SRS calibration variables. When possible, a weight of 1 was used as the lower limit of the weights for these geography levels. If calibration was not possible with a minimum weight of 1, lower minimum weights were attempted, starting at 0.998 and then decreasing from 0.99 to 0.9 in increments of 0.01 until calibration was successful. Thus any weights < 0.9 are illogical and should be investigated. Note that calibration success is affected by parameters supplied to the gencalib() function (e.g., design weights and upper bound of g-weights). If weights of at least 0.9 and less than 1 occur, it is possible that a solution with a minimum weight of 1 exists but was not found in the given number of iterations. 
* In the state-, judicial district-, MSA-, and field office-level tables, weights less than 1 are marked red. During calibration of state-, judicial district-, MSA-, and field office-level weights, a weight of 1 was used as the lower limit thus any weights less than this number suggest the model failed for a weighting group. If weights < 1 do occur, review the domain where it occurs to ensure the process is working correctly. 
* In the university and tribal tables, weights less than 1 are marked red. Weights there were assigned by N/n, where N is the number of NIBRS-eligible agencies and n is the number of NIBRS-reporting agencies; weights < 1 is theoretically illogical and suggest an issue with either the numerator or denominator of the weights calculation. If weights < 1 do occur, review the domain where it occurs to ensure the process is working correctly. 
* UWEs between 2 and 3 are marked yellow in all the tables and UWEs greater than 3 are marked red. This indicates a high level of variance inflation due to unequal weighting. For national-, region-, state-, judicial district-, MSA-, and field office-level weights this implies that the agencies within a weighting group had high levels of dispersion in their weights. If this occurs in a domain, review the coverage rate for the domain. If the coverage is low, this result is expected and not an indication of a larger issue.
* Cases where the 95th percentile is 5 or more times higher than the median are marked red in the national-, region-, state-, judicial district-, MSA-, field office-level weights. For these weight variables, this implies one or more agencies had significantly higher weights than the rest of the agencies in a given summary level. All tribal and university LEAs received the same weight and thus this isn’t applicable in the corresponding tables. If this occurs in a domain, review the coverage rate for the domain. If the coverage is low, this result is possible and not an indication of a larger issue.


## Load Data

Loading national, region, state, university, and tribal checks

```{r load,echo=FALSE,message=FALSE,include=FALSE}

log_trace("Generating weights_checks.html")
#################
#Read in files
results_national <- read_csv(file=paste0(output_weighting_data_folder,"weights_national_checks.csv"))
results_region <- read_csv(file=paste0(output_weighting_data_folder,"weights_region_checks.csv"))
results_state <- read_csv(file=paste0(output_weighting_data_folder,"weights_state_checks.csv"))
results_jd <- read_csv(file=paste0(output_weighting_data_folder,"weights_jd_cal_srs_altcombs_col_checks.csv"))
results_msa <- read_csv(file=paste0(output_weighting_data_folder,"weights_MSA_cal_srs_altcombs_col_checks.csv"))
results_fo <- read_csv(file=paste0(output_weighting_data_folder,"weights_FO_cal_srs_altcombs_col_checks.csv"))
results_university <- read_csv(file=paste0(output_weighting_data_folder,"weights_university_checks.csv"))
results_tribal <- read_csv(file=paste0(output_weighting_data_folder,"weights_tribal_checks.csv"))
#27Jul2022: Read in national weights
weights_national <- read_csv(paste0(output_weighting_data_folder,'weights_national.csv'))

```

## Prep overall national results
```{r prep,echo=FALSE,message=FALSE,include=FALSE}

SF2 <- weights_national
out_temp <- SF2 %>%
  subset(!is.na(NationalWgt))
temp.describe <- describe(out_temp$NationalWgt)
temp.quantiles <- quantile(out_temp$NationalWgt,
                           probs=c(0,0.05,0.1,0.25,0.5,0.75,0.9,0.95,1),
                           na.rm=TRUE)
#Note (26Jul2022): Adding n Eligible LEAs
temp.nElig <- SF2 %>%
  nrow()
#Note (16May2024): round weights to 6 digits before checking if <1
temp.nLT1 <- sum(round(out_temp$NationalWgt,digits=6) < 1 & out_temp$NationalWgt>0, na.rm=TRUE)
#Note (14May2025): check for weights <0.9
temp.nLT0pt9 <- sum(round(out_temp$NationalWgt,digits=6) < 0.9 & out_temp$NationalWgt>0, na.rm=TRUE)
#Note (29Jul2022): Switching from >100 to >20
#temp.nGT100 <- sum(out_temp$NationalWgt > 100, na.rm=TRUE)
temp.nGT20 <- sum(out_temp$NationalWgt > 20, na.rm=TRUE)
temp.UWE <- 1+var(out_temp$NationalWgt,na.rm=TRUE)/(mean(out_temp$NationalWgt,na.rm=TRUE)^2)
#Note (16May2024): adding min and max
temp.out <- data.frame(wgtGpDesc="All LEAs",
                       calibrated=NA,
                       nElig=temp.nElig,
                       nNIBRS=as.numeric(temp.describe$counts["n"]) + as.numeric(temp.describe$counts["missing"]),
                       nMissing=as.numeric(temp.describe$counts["missing"]),
                       nLT1=temp.nLT1,
                       nLT0pt9=temp.nLT0pt9,
                       #nGT100=temp.nGT100,
                       nGT20=temp.nGT20,
                       UWE=sprintf(temp.UWE,fmt="%1.3f"),
                       Mean=sprintf(as.numeric(temp.describe$counts["Mean"]),fmt="%1.3f"),
                       Min=sprintf(temp.quantiles["0%"],fmt="%1.3f"),
                       pt05=sprintf(temp.quantiles["5%"],fmt="%1.3f"),
                       pt10=sprintf(temp.quantiles["10%"],fmt="%1.3f"),
                       pt25=sprintf(temp.quantiles["25%"],fmt="%1.3f"),
                       pt50=sprintf(temp.quantiles["50%"],fmt="%1.3f"),
                       pt75=sprintf(temp.quantiles["75%"],fmt="%1.3f"),
                       pt90=sprintf(temp.quantiles["90%"],fmt="%1.3f"),
                       pt95=sprintf(temp.quantiles["95%"],fmt="%1.3f"),
                       Max=sprintf(temp.quantiles["100%"],fmt="%1.3f")) %>%
  #Need to convert some variables to numeric
  mutate(UWE=as.numeric(UWE),
         Mean=as.numeric(Mean),
         Min=as.numeric(Min),
         pt05=as.numeric(pt05),
         pt10=as.numeric(pt10),
         pt25=as.numeric(pt25),
         pt50=as.numeric(pt50),
         pt75=as.numeric(pt75),
         pt90=as.numeric(pt90),
         pt95=as.numeric(pt95),
         Max=as.numeric(Max))
colnames(temp.out) <- c("Weight Group","Calibrated",
                        "n Eligible LEAs","n NIBRS LEAs, Overall","n NIBRS LEAs, Missing Weight",
                        "n LT 1","n LT 0.9","n GT 20","UWE",#"n GT 100","UWE",
                        "Mean","Minimum","5th Pctl","10th Pctl","25th Pctl","50th Pctl",
                        "75th Pctl","90th Pctl","95th Pctl","Maximum") 
list2env(list("results_national_0"=temp.out),.GlobalEnv)

#Stack total row onto rest of national results
results_national <- bind_rows(results_national_0,
                              results_national)
```

## Display Results
```{r results,echo=FALSE,warning=FALSE}
#################
#Display results

#29Jul2022: Changing n GT 100 to n GT 20

#30Jun2023: Adding summary table before each geography


#National
message("National")

#14May2025: dropping weights <1 from major to minor issue, and adding weights <0.9 as major issue

summary_national_major <- results_national %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(redMissingWgts=`n NIBRS LEAs, Missing Weight`>0,
         yellowNLT1=`n LT 1`>0,
         redNLT0pt9=`n LT 0.9`>0,
         redNGT20=`n GT 20`>0,
         yellowUWE=UWE<3 & UWE>=2,
         redUWE=UWE>=3,
         redRatio=ratio>=5) %>%
  dplyr::summarize(anyMajor=any(redMissingWgts,redNLT0pt9,redNGT20,redUWE,redRatio),
                   anyMissingWgts=any(redMissingWgts),
                   anyMissingWgtsGeo=ifelse(anyMissingWgts,"National",NA_character_),
                   anyNLT0pt9=any(redNLT0pt9),
                   anyNLT0pt9Geo=ifelse(anyNLT0pt9,"National",NA_character_),
                   anyNGT20=any(redNGT20),
                   anyNGT20Geo=ifelse(anyNGT20,"National",NA_character_),
                   anyUWEGTE3=any(redUWE),
                   anyUWEGTE3Geo=ifelse(anyUWEGTE3,"National",NA_character_),
                   anyRatioGTE5=any(redRatio),
                   anyRatioGTE5Geo=ifelse(anyRatioGTE5,"National",NA_character_),
                   anyMissingWgtsDesc=ifelse(any(anyMissingWgts),paste0("1+ NIBRS LEAs missing weights",str_flatten(anyMissingWgtsGeo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNLT0pt9Desc=ifelse(any(anyNLT0pt9),paste0("1+ weight < 0.9: ",str_flatten(anyNLT0pt9Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNGT20Desc=ifelse(any(anyNGT20),paste0("1+ weight > 20: ",str_flatten(anyNGT20Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyUWEGTE3Desc=ifelse(any(anyUWEGTE3),paste0("UWE >= 3: ",str_flatten(anyUWEGTE3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyRatioGTE5Desc=ifelse(any(anyRatioGTE5),paste0("95th Pctl / Median >= 5: ",str_flatten(anyRatioGTE5Geo %>% subset(!is.na(.)),col=", ")),NA_character_)) %>%
  mutate(anyMajorDesc=ifelse(any(anyMajor),
                             str_flatten(c(anyMissingWgtsDesc %>% subset(!is.na(.)),
                                           anyNLT0pt9Desc %>% subset(!is.na(.)),
                                           anyNGT20Desc %>% subset(!is.na(.)),
                                           anyUWEGTE3Desc %>% subset(!is.na(.)),
                                           anyRatioGTE5Desc %>% subset(!is.na(.))),
                                         col=";<br/><br/>"),
                             NA_character_)) %>%
  mutate(`Weight Level`="National",
         Status=ifelse(anyMajor,"Major problem(s) found","No major problems found"),
         `Flagged Areas`=anyMajorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_national_minor <- results_national %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(yellowNLT1=`n LT 1`>0,
         yellowUWE=UWE<3 & UWE>=2) %>%
  dplyr::summarize(anyMinor=any(yellowNLT1,yellowUWE),
                   anyNLT1=any(yellowNLT1),
                   anyNLT1Desc=ifelse(anyNLT1,"1+ weight <1",NA_character_),
                   anyUWE2to3=any(yellowUWE),
                   anyUWE2to3Desc=ifelse(anyUWE2to3,"2 <= UWE < 3",NA_character_),
                   anyMinorDesc=ifelse(any(anyMinor),
                                       str_flatten(c(anyNLT1Desc %>% subset(!is.na(.)),
                                                     anyUWE2to3Desc %>% subset(!is.na(.))),
                                                   col=";<br/><br/>"),
                                       NA_character_)) %>% 
  mutate(`Weight Level`="National",
         Status=ifelse(anyMinor,"Minor problem(s) found","No minor problems found"),
         `Flagged Areas`=anyMinorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_national <- bind_rows(summary_national_major,
                              summary_national_minor)

summary_national %>% DT::datatable()

results_national %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  `row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = 20)#17)#16)
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','yellow'))) %>%
  formatStyle(columns="n LT 0.9",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) %>%
  formatStyle(columns="95th Pctl",
              valueColumns="ratio",
              backgroundColor = styleInterval(c(5), c('NULL','red')))

#Region
message("Region")

#14May2025: dropping weights <1 from major to minor issue, and adding weights <0.9 as major issue

summary_region_major <- results_region %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(redMissingWgts=`n NIBRS LEAs, Missing Weight`>0,
         yellowNLT1=`n LT 1`>0,
         redNLT0pt9=`n LT 0.9`>0,
         redNGT20=`n GT 20`>0,
         yellowUWE=UWE<3 & UWE>=2,
         redUWE=UWE>=3,
         redRatio=ratio>=5) %>%
  group_by(`Region`) %>%
  dplyr::summarize(anyMajor=any(redMissingWgts,redNLT0pt9,redNGT20,redUWE,redRatio),
                   anyMissingWgts=any(redMissingWgts),
                   anyMissingWgtsGeo=ifelse(anyMissingWgts,`Region`,NA_character_),
                   anyNLT0pt9=any(redNLT0pt9),
                   anyNLT0pt9Geo=ifelse(anyNLT0pt9,`Region`,NA_character_),
                   anyNGT20=any(redNGT20),
                   anyNGT20Geo=ifelse(anyNGT20,`Region`,NA_character_),
                   anyUWEGTE3=any(redUWE),
                   anyUWEGTE3Geo=ifelse(anyUWEGTE3,`Region`,NA_character_),
                   anyRatioGTE5=any(redRatio),
                   anyRatioGTE5Geo=ifelse(anyRatioGTE5,`Region`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyMajor=any(anyMajor),
                   anyMissingWgtsDesc=ifelse(any(anyMissingWgts),paste0("1+ NIBRS LEAs missing weights",str_flatten(anyMissingWgtsGeo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNLT0pt9Desc=ifelse(any(anyNLT0pt9),paste0("1+ weight < 0.9: ",str_flatten(anyNLT0pt9Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNGT20Desc=ifelse(any(anyNGT20),paste0("1+ weight > 20: ",str_flatten(anyNGT20Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyUWEGTE3Desc=ifelse(any(anyUWEGTE3),paste0("UWE >= 3: ",str_flatten(anyUWEGTE3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyRatioGTE5Desc=ifelse(any(anyRatioGTE5),paste0("95th Pctl / Median >= 5: ",str_flatten(anyRatioGTE5Geo %>% subset(!is.na(.)),col=", ")),NA_character_)) %>%
  mutate(anyMajorDesc=ifelse(any(anyMajor),
                             str_flatten(c(anyMissingWgtsDesc %>% subset(!is.na(.)),
                                           anyNLT0pt9Desc %>% subset(!is.na(.)),
                                           anyNGT20Desc %>% subset(!is.na(.)),
                                           anyUWEGTE3Desc %>% subset(!is.na(.)),
                                           anyRatioGTE5Desc %>% subset(!is.na(.))),
                                         col=";<br/><br/>"),
                             NA_character_)) %>%
  mutate(`Weight Level`="Region",
         Status=ifelse(!is.na(anyMajorDesc),"Major problem(s) found","No major problems found"),
         `Flagged Areas`=anyMajorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_region_minor <- results_region %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(yellowNLT1=`n LT 1`>0,
         yellowUWE=UWE<3 & UWE>=2) %>%
  group_by(Region) %>%
  dplyr::summarize(anyMinor=any(yellowNLT1,yellowUWE),
                   anyNLT1=any(yellowNLT1),
                   anyNLT1Geo=ifelse(anyNLT1,`Region`,NA_character_),
                   anyUWE2to3=any(yellowUWE),
                   anyUWE2to3Geo=ifelse(anyUWE2to3,`Region`,NA_character_)) %>% 
  dplyr::summarize(anyMinor=any(anyMinor),
                   anyNLT1Desc=ifelse(any(anyNLT1),paste0("1+ weight <1: ",str_flatten(anyNLT1Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyUWE2to3Desc=ifelse(any(anyUWE2to3),paste0("2 <= UWE < 3: ",str_flatten(anyUWE2to3Geo %>% subset(!is.na(.)),col=", ")),NA_character_)) %>%
  mutate(anyMinorDesc=ifelse(any(anyMinor),
                             str_flatten(c(anyNLT1Desc %>% subset(!is.na(.)),
                                           anyUWE2to3Desc %>% subset(!is.na(.))),
                                         col=";<br/><br/>"),
                             NA_character_)) %>%
  mutate(`Weight Level`="Region",
         Status=ifelse(!is.na(anyMinorDesc),"Minor problem(s) found","No minor problems found"),
         `Flagged Areas`=anyMinorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_region <- bind_rows(summary_region_major,
                            summary_region_minor)

summary_region %>% DT::datatable(escape=FALSE)

results_region %>%
  arrange(Region) %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  `row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = 21)#18)#17)
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','yellow'))) %>%
  formatStyle(columns="n LT 0.9",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) %>%
  formatStyle(columns="95th Pctl",
              valueColumns="ratio",
              backgroundColor = styleInterval(c(5), c('NULL','red')))

#State
message("State")
summary_state_major <- results_state %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(redMissingWgts=`n NIBRS LEAs, Missing Weight`>0,
         redNLT1=`n LT 1`>0,
         redNGT20=`n GT 20`>0,
         yellowUWE=UWE<3 & UWE>=2,
         redUWE=UWE>=3,
         redRatio=ratio>=5) %>%
  group_by(`State`) %>%
  dplyr::summarize(anyMajor=any(redMissingWgts,redNLT1,redNGT20,redUWE,redRatio),
                   anyMissingWgts=any(redMissingWgts),
                   anyMissingWgtsGeo=ifelse(anyMissingWgts,`State`,NA_character_),
                   anyNLT1=any(redNLT1),
                   anyNLT1Geo=ifelse(anyNLT1,`State`,NA_character_),
                   anyNGT20=any(redNGT20),
                   anyNGT20Geo=ifelse(anyNGT20,`State`,NA_character_),
                   anyUWEGTE3=any(redUWE),
                   anyUWEGTE3Geo=ifelse(anyUWEGTE3,`State`,NA_character_),
                   anyRatioGTE5=any(redRatio),
                   anyRatioGTE5Geo=ifelse(anyRatioGTE5,`State`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyMajor=any(anyMajor),
                   anyMissingWgtsDesc=ifelse(any(anyMissingWgts),paste0("1+ NIBRS LEAs missing weights",str_flatten(anyMissingWgtsGeo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNLT1Desc=ifelse(any(anyNLT1),paste0("1+ weight <1: ",str_flatten(anyNLT1Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNGT20Desc=ifelse(any(anyNGT20),paste0("1+ weight > 20: ",str_flatten(anyNGT20Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyUWEGTE3Desc=ifelse(any(anyUWEGTE3),paste0("UWE >= 3: ",str_flatten(anyUWEGTE3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyRatioGTE5Desc=ifelse(any(anyRatioGTE5),paste0("95th Pctl / Median >= 5: ",str_flatten(anyRatioGTE5Geo %>% subset(!is.na(.)),col=", ")),NA_character_)) %>%
  mutate(anyMajorDesc=ifelse(any(anyMajor),
                             str_flatten(c(anyMissingWgtsDesc %>% subset(!is.na(.)),
                                           anyNLT1Desc %>% subset(!is.na(.)),
                                           anyNGT20Desc %>% subset(!is.na(.)),
                                           anyUWEGTE3Desc %>% subset(!is.na(.)),
                                           anyRatioGTE5Desc %>% subset(!is.na(.))),
                                         col=";<br/><br/>"),
                             NA_character_)) %>%
  mutate(`Weight Level`="State",
         Status=ifelse(!is.na(anyMajorDesc),"Major problem(s) found","No major problems found"),
         `Flagged Areas`=anyMajorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_state_minor <- results_state %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(yellowUWE=UWE<3 & UWE>=2) %>%
  group_by(`State`) %>%
  dplyr::summarize(anyMinor=any(yellowUWE),
                   anyUWE2to3=any(yellowUWE),
                   anyUWE2to3Geo=ifelse(anyUWE2to3,`State`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyUWE2to3Desc=ifelse(any(anyUWE2to3),paste0("2 <= UWE < 3: ",str_flatten(anyUWE2to3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyMinor=any(anyUWE2to3)) %>%
  mutate(anyMinorDesc=ifelse(anyMinor,anyUWE2to3Desc,NA_character_)) %>%
  mutate(`Weight Level`="State",
         Status=ifelse(!is.na(anyMinorDesc),"Minor problem(s) found","No minor problems found"),
         `Flagged Areas`=anyMinorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_state <- bind_rows(summary_state_major,
                           summary_state_minor)

summary_state %>% DT::datatable(escape=FALSE)

results_state %>%
  arrange(State) %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  `row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = 20)#16)#15)
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) %>%
  formatStyle(columns="95th Pctl",
              valueColumns="ratio",
              backgroundColor = styleInterval(c(5), c('NULL','red')))


#JD
message("JD")

summary_jd_major <- results_jd %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(redMissingWgts=`n NIBRS LEAs, Missing Weight`>0,
         redNLT1=`n LT 1`>0,
         redNGT20=`n GT 20`>0,
         yellowUWE=UWE<3 & UWE>=2,
         redUWE=UWE>=3,
         redRatio=ratio>=5) %>%
  group_by(`Judicial District`) %>%
  dplyr::summarize(anyMajor=any(redMissingWgts,redNLT1,redNGT20,redUWE,redRatio),
                   anyMissingWgts=any(redMissingWgts),
                   anyMissingWgtsGeo=ifelse(anyMissingWgts,`Judicial District`,NA_character_),
                   anyNLT1=any(redNLT1),
                   anyNLT1Geo=ifelse(anyNLT1,`Judicial District`,NA_character_),
                   anyNGT20=any(redNGT20),
                   anyNGT20Geo=ifelse(anyNGT20,`Judicial District`,NA_character_),
                   anyUWEGTE3=any(redUWE),
                   anyUWEGTE3Geo=ifelse(anyUWEGTE3,`Judicial District`,NA_character_),
                   anyRatioGTE5=any(redRatio),
                   anyRatioGTE5Geo=ifelse(anyRatioGTE5,`Judicial District`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyMajor=any(anyMajor),
                   anyMissingWgtsDesc=ifelse(any(anyMissingWgts),paste0("1+ NIBRS LEAs missing weights",str_flatten(anyMissingWgtsGeo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNLT1Desc=ifelse(any(anyNLT1),paste0("1+ weight <1: ",str_flatten(anyNLT1Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNGT20Desc=ifelse(any(anyNGT20),paste0("1+ weight > 20: ",str_flatten(anyNGT20Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyUWEGTE3Desc=ifelse(any(anyUWEGTE3),paste0("UWE >= 3: ",str_flatten(anyUWEGTE3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyRatioGTE5Desc=ifelse(any(anyRatioGTE5),paste0("95th Pctl / Median >= 5: ",str_flatten(anyRatioGTE5Geo %>% subset(!is.na(.)),col=", ")),NA_character_)) %>%
  mutate(anyMajorDesc=ifelse(any(anyMajor),
                             str_flatten(c(anyMissingWgtsDesc %>% subset(!is.na(.)),
                                           anyNLT1Desc %>% subset(!is.na(.)),
                                           anyNGT20Desc %>% subset(!is.na(.)),
                                           anyUWEGTE3Desc %>% subset(!is.na(.)),
                                           anyRatioGTE5Desc %>% subset(!is.na(.))),
                                         col=";<br/><br/>"),
                             NA_character_)) %>%
  mutate(`Weight Level`="Judicial District",
         Status=ifelse(!is.na(anyMajorDesc),"Major problem(s) found","No major problems found"),
         `Flagged Areas`=anyMajorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_jd_minor <- results_jd %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(yellowUWE=UWE<3 & UWE>=2) %>%
  group_by(`Judicial District`) %>%
  dplyr::summarize(anyMinor=any(yellowUWE),
                   anyUWE2to3=any(yellowUWE),
                   anyUWE2to3Geo=ifelse(anyUWE2to3,`Judicial District`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyUWE2to3Desc=ifelse(any(anyUWE2to3),paste0("2 <= UWE < 3: ",str_flatten(anyUWE2to3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyMinor=any(anyUWE2to3)) %>%
  mutate(anyMinorDesc=ifelse(anyMinor,anyUWE2to3Desc,NA_character_)) %>%
  mutate(`Weight Level`="Judicial District",
         Status=ifelse(!is.na(anyMinorDesc),"Minor problem(s) found","No minor problems found"),
         `Flagged Areas`=anyMinorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_jd <- bind_rows(summary_jd_major,
                        summary_jd_minor)

summary_jd %>% DT::datatable(escape=FALSE)

results_jd %>%
  arrange(`Judicial District`) %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  #`row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = 20)#18)#16)#15)
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) %>%
  formatStyle(columns="95th Pctl",
              valueColumns="ratio",
              backgroundColor = styleInterval(c(5), c('NULL','red')))

#MSA
message("MSA")
summary_msa_major <- results_msa %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(redMissingWgts=`n NIBRS LEAs, Missing Weight`>0,
         redNLT1=`n LT 1`>0,
         redNGT20=`n GT 20`>0,
         yellowUWE=UWE<3 & UWE>=2,
         redUWE=UWE>=3,
         redRatio=ratio>=5) %>%
  group_by(`MSA`) %>%
  dplyr::summarize(anyMajor=any(redMissingWgts,redNLT1,redNGT20,redUWE,redRatio),
                   anyMissingWgts=any(redMissingWgts),
                   anyMissingWgtsGeo=ifelse(anyMissingWgts,`MSA`,NA_character_),
                   anyNLT1=any(redNLT1),
                   anyNLT1Geo=ifelse(anyNLT1,`MSA`,NA_character_),
                   anyNGT20=any(redNGT20),
                   anyNGT20Geo=ifelse(anyNGT20,`MSA`,NA_character_),
                   anyUWEGTE3=any(redUWE),
                   anyUWEGTE3Geo=ifelse(anyUWEGTE3,`MSA`,NA_character_),
                   anyRatioGTE5=any(redRatio),
                   anyRatioGTE5Geo=ifelse(anyRatioGTE5,`MSA`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyMajor=any(anyMajor),
                   anyMissingWgtsDesc=ifelse(any(anyMissingWgts),paste0("1+ NIBRS LEAs missing weights",str_flatten(anyMissingWgtsGeo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNLT1Desc=ifelse(any(anyNLT1),paste0("1+ weight <1: ",str_flatten(anyNLT1Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNGT20Desc=ifelse(any(anyNGT20),paste0("1+ weight > 20: ",str_flatten(anyNGT20Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyUWEGTE3Desc=ifelse(any(anyUWEGTE3),paste0("UWE >= 3: ",str_flatten(anyUWEGTE3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyRatioGTE5Desc=ifelse(any(anyRatioGTE5),paste0("95th Pctl / Median >= 5: ",str_flatten(anyRatioGTE5Geo %>% subset(!is.na(.)),col=", ")),NA_character_)) %>%
  mutate(anyMajorDesc=ifelse(any(anyMajor),
                             str_flatten(c(anyMissingWgtsDesc %>% subset(!is.na(.)),
                                           anyNLT1Desc %>% subset(!is.na(.)),
                                           anyNGT20Desc %>% subset(!is.na(.)),
                                           anyUWEGTE3Desc %>% subset(!is.na(.)),
                                           anyRatioGTE5Desc %>% subset(!is.na(.))),
                                         col=";<br/><br/>"),
                             NA_character_)) %>%
  mutate(`Weight Level`="MSA",
         Status=ifelse(!is.na(anyMajorDesc),"Major problem(s) found","No major problems found"),
         `Flagged Areas`=anyMajorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_msa_minor <- results_msa %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(yellowUWE=UWE<3 & UWE>=2) %>%
  group_by(`MSA`) %>%
  dplyr::summarize(anyMinor=any(yellowUWE),
                   anyUWE2to3=any(yellowUWE),
                   anyUWE2to3Geo=ifelse(anyUWE2to3,`MSA`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyUWE2to3Desc=ifelse(any(anyUWE2to3),paste0("2 <= UWE < 3: ",str_flatten(anyUWE2to3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyMinor=any(anyUWE2to3)) %>%
  mutate(anyMinorDesc=ifelse(anyMinor,anyUWE2to3Desc,NA_character_)) %>%
  mutate(`Weight Level`="MSA",
         Status=ifelse(!is.na(anyMinorDesc),"Minor problem(s) found","No minor problems found"),
         `Flagged Areas`=anyMinorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_msa <- bind_rows(summary_msa_major,
                         summary_msa_minor)

summary_msa %>% DT::datatable(escape=FALSE)

results_msa %>%
  arrange(MSA) %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  #`row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = 20)#18)#16)#15)
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) %>%
  formatStyle(columns="95th Pctl",
              valueColumns="ratio",
              backgroundColor = styleInterval(c(5), c('NULL','red')))

#FO
message("FO")

summary_fo_major <- results_fo %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(redMissingWgts=`n NIBRS LEAs, Missing Weight`>0,
         redNLT1=`n LT 1`>0,
         redNGT20=`n GT 20`>0,
         yellowUWE=UWE<3 & UWE>=2,
         redUWE=UWE>=3,
         redRatio=ratio>=5) %>%
  group_by(`Field Office`) %>%
  dplyr::summarize(anyMajor=any(redMissingWgts,redNLT1,redNGT20,redUWE,redRatio),
                   anyMissingWgts=any(redMissingWgts),
                   anyMissingWgtsGeo=ifelse(anyMissingWgts,`Field Office`,NA_character_),
                   anyNLT1=any(redNLT1),
                   anyNLT1Geo=ifelse(anyNLT1,`Field Office`,NA_character_),
                   anyNGT20=any(redNGT20),
                   anyNGT20Geo=ifelse(anyNGT20,`Field Office`,NA_character_),
                   anyUWEGTE3=any(redUWE),
                   anyUWEGTE3Geo=ifelse(anyUWEGTE3,`Field Office`,NA_character_),
                   anyRatioGTE5=any(redRatio),
                   anyRatioGTE5Geo=ifelse(anyRatioGTE5,`Field Office`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyMajor=any(anyMajor),
                   anyMissingWgtsDesc=ifelse(any(anyMissingWgts),paste0("1+ NIBRS LEAs missing weights",str_flatten(anyMissingWgtsGeo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNLT1Desc=ifelse(any(anyNLT1),paste0("1+ weight <1: ",str_flatten(anyNLT1Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyNGT20Desc=ifelse(any(anyNGT20),paste0("1+ weight > 20: ",str_flatten(anyNGT20Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyUWEGTE3Desc=ifelse(any(anyUWEGTE3),paste0("UWE >= 3: ",str_flatten(anyUWEGTE3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyRatioGTE5Desc=ifelse(any(anyRatioGTE5),paste0("95th Pctl / Median >= 5: ",str_flatten(anyRatioGTE5Geo %>% subset(!is.na(.)),col=", ")),NA_character_)) %>%
  mutate(anyMajorDesc=ifelse(any(anyMajor),
                             str_flatten(c(anyMissingWgtsDesc %>% subset(!is.na(.)),
                                           anyNLT1Desc %>% subset(!is.na(.)),
                                           anyNGT20Desc %>% subset(!is.na(.)),
                                           anyUWEGTE3Desc %>% subset(!is.na(.)),
                                           anyRatioGTE5Desc %>% subset(!is.na(.))),
                                         col=";<br/><br/>"),
                             NA_character_)) %>%
  mutate(`Weight Level`="Field Office",
         Status=ifelse(!is.na(anyMajorDesc),"Major problem(s) found","No major problems found"),
         `Flagged Areas`=anyMajorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_fo_minor <- results_fo %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  mutate(yellowUWE=UWE<3 & UWE>=2) %>%
  group_by(`Field Office`) %>%
  dplyr::summarize(anyMinor=any(yellowUWE),
                   anyUWE2to3=any(yellowUWE),
                   anyUWE2to3Geo=ifelse(anyUWE2to3,`Field Office`,NA_character_)) %>%
  ungroup() %>%
  dplyr::summarize(anyUWE2to3Desc=ifelse(any(anyUWE2to3),paste0("2 <= UWE < 3: ",str_flatten(anyUWE2to3Geo %>% subset(!is.na(.)),col=", ")),NA_character_),
                   anyMinor=any(anyUWE2to3)) %>%
  mutate(anyMinorDesc=ifelse(anyMinor,anyUWE2to3Desc,NA_character_)) %>%
  mutate(`Weight Level`="Field Office",
         Status=ifelse(!is.na(anyMinorDesc),"Minor problem(s) found","No minor problems found"),
         `Flagged Areas`=anyMinorDesc) %>%
  select(`Weight Level`,Status,`Flagged Areas`)

summary_fo <- bind_rows(summary_fo_major,
                        summary_fo_minor)

summary_fo %>% DT::datatable(escape=FALSE)

results_fo %>%
  arrange(`Field Office`) %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  #`row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = 20)#18)#16)#15)
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) %>%
  formatStyle(columns="95th Pctl",
              valueColumns="ratio",
              backgroundColor = styleInterval(c(5), c('NULL','red')))

#University
message("University")
#Note (12Jul2022): Don't show percentiles and add note
results_university %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  #`row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = c(9:18))#c(9:16))#targets = 15
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) #%>%
#formatStyle(columns="95th Pctl",
#		  valueColumns="ratio",
#          backgroundColor = styleInterval(c(5), c('NULL','red')))
message("All agencies in this agency type have the same weight so percentiles are not displayed")

#Tribal
message("Tribal")
#Note (12Jul2022): Don't show percentiles and add note
results_tribal %>%
  mutate(ratio=as.numeric(eval(as.symbol("95th Pctl")))/as.numeric(eval(as.symbol("50th Pctl")))) %>%
  #`row.names<-`(1:nrow(.)) %>%
  DT::datatable(options = list(columnDefs = list(
    list(visible = FALSE, targets = c(9:18))#c(9:16))#targets = 15
  ))) %>%
  formatStyle(columns="n NIBRS LEAs, Missing Weight",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="n LT 1",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  #formatStyle(columns="n GT 100",
  formatStyle(columns="n GT 20",
              backgroundColor = styleInterval(c(0), c('NULL','red'))) %>%
  formatStyle(columns="UWE",
              backgroundColor = styleInterval(c(2,3), c('NULL','yellow','red'))) #%>%
#formatStyle(columns="95th Pctl",
#			  valueColumns="ratio",
#             backgroundColor = styleInterval(c(5), c('NULL','red')))
message("All agencies in this agency type have the same weight so percentiles are not displayed")
log_trace("Finished generating weights_checks.html")
```
